requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.2.2
Django==2.0.3
django-filter==1.1.0
djangorestframework==3.8.2
dwebsocket==0.4.2
gevent==1.2.2
greenlet==0.4.13
httpie==0.9.9
idna==2.6
Pillow==5.1.0
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
pyparsing==2.2.0
pytz==2018.3
requests==2.18.4
six==1.11.0
svgwrite==1.1.12
Tree==0.2.3
urllib3==1.22
